This is the first commit for this repository
My team and I wrote this project in the fall semester of 2022, for the final project for one course. We used HTML, CSS, SASS with programming languages JavaScript, PHP. We used JavaScript for user interaction and PHP for a database to store SQL data. The screen size of our project is adapted to three categories: laptop, iPad Air, Galaxy A71.

The project consists of many pages, all the necessary materials, such as photos and logos, were taken from the original web page - (turanwater.com) . We did not copy the code from any source, this project was written in order to strengthen our knowledge in the field of web development. We respect the owners of this site and are not responsible for any violations of rights. All actions were carried out under the guidance of our university to assess our knowledge.